+++ NordicDoorSuggestionSystem/Entities/Team.cs
@@ -11,7 +11,7 @@ namespace NordicDoorSuggestionSystem.Entities
         public string? TeamName { get; set; }
         [MaxLength(50)]
         public string? TeamLeader { get; set; }
-        public ushort? TeamSgstnCount { get; set; }
+        public int? TeamSgstnCount { get; set; }
         
         public Department? Department { get; set; }
         [ForeignKey("Department")]
